restructured folders/file names
